breakout some logic from the current step route
