+++ apps/newsletters-api/src/app/state-machine.ts
@@ -1,0 +1,55 @@
+import { newslettersWorkflowStepLayout } from '@newsletters-nx/newsletter-workflow';
+import type {
+	CurrentStepRouteRequest,
+	CurrentStepRouteResponse,
+	WizardButton,
+	WizardStepData,
+	WizardStepLayout,
+	WizardStepLayoutButton,
+} from '@newsletters-nx/state-machine';
+import {
+	setupInitialState,
+	stateMachineButtonPressed,
+} from '@newsletters-nx/state-machine';
+import { storageInstance } from '../services/storageInstance';
+
+export const convertWizardStepLayoutButtonsToWizardButtons = (
+	layoutButtons: WizardStepLayout['buttons'],
+): CurrentStepRouteResponse['buttons'] => {
+	const convertButton = (
+		index: string,
+		input: WizardStepLayoutButton,
+	): WizardButton => {
+		return {
+			id: index,
+			label: input.label,
+			buttonType: input.buttonType,
+		};
+	};
+
+	const outputRecord: CurrentStepRouteResponse['buttons'] = {};
+
+	Object.entries(layoutButtons).forEach(([index, layoutButton]) => {
+		outputRecord[index] = convertButton(index, layoutButton);
+	});
+
+	return outputRecord;
+};
+
+export const unsafelyGetState = async (
+	body: CurrentStepRouteRequest,
+): Promise<WizardStepData> => {
+	if (body.buttonId !== undefined) {
+		return await stateMachineButtonPressed(
+			body.buttonId,
+			{
+				currentStepId: body.stepId,
+				formData: body.formData,
+			},
+			newslettersWorkflowStepLayout,
+			storageInstance,
+		);
+	} else {
+		return await setupInitialState(body, storageInstance);
+	}
+};
